Refactor throttler module
